Timing changes - Check if current time is valid for bet - Reorganized time funciton dependency between model and match_getter
